% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formula.R
\name{create_formula}
\alias{create_formula}
\title{Create a sgb formula}
\usage{
create_formula(
  alpha = 0.05,
  group_df = NULL,
  blearner = "bols",
  outcome_name = "y",
  group_name = "group_name",
  var_name = "var_name"
)
}
\arguments{
\item{alpha}{mixing parameter}

\item{group_df}{data frame containing variable names with group structure}

\item{blearner}{type of baselearner}

\item{outcome_name}{name of dependent variable}

\item{group_name}{name of column in group_df indicating the group structure of the variables}

\item{var_name}{name of column in group_df containing the variable names}
}
\value{
formula to be passed to mboost() yielding the sparse group boosting for a given value mixing parameter alpha.
}
\description{
Create a sgb formula
}
