% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varimp.R
\name{get_varimp}
\alias{get_varimp}
\title{Create a sparse-group boosting formula}
\usage{
get_varimp(sgb_model)
}
\arguments{
\item{model}{mboost model to compute the variable importance for.}
}
\value{
dataframe containing the variables, group structure and
variable importance on both group and individual variable basis.
}
\description{
Commutes the variable importance of a sparse-group mboost model.
}
\examples{
library(mboost)
library(tidyverse)
set.seed(1)
df <- data.frame(
 x1 = rnorm(100),x2 = rnorm(100),x3 = rnorm(100),
 x4 = rnorm(100), x5 = runif(100)
 )
df <- df \%>\%
mutate_all(function(x){as.numeric(scale(x))})
df$y <- df$x1+df$x4+df$x5
group_df <- data.frame(
 group_name = c(1,1,1,2,2),
 var_name = c('x1','x2','x3','x4','x5')
)

sgb_formula <- create_formula(alpha = 0.3, group_df = group_df)
sgb_model <- mboost(formula = sgb_formula, data = df)
sgb_varimp <- get_varimp(sgb_model)
}
