% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_effects}
\alias{plot_effects}
\title{Variable importance of a sparse group boosting model}
\usage{
plot_effects(sgb_model, prop = 1, n_vars = 30, max_char_length = 5)
}
\arguments{
\item{sgb_model}{mboost model to compute the variable importance for.}

\item{prop}{the maximum proportion of explained importance. Default value is one,
meaning all predictors are plotted. By setting prop smaller than one the number of
plotted variables can be reduced. One can also use 'nvars' for limiting
the number of variables to be plotted directly.}

\item{n_vars}{The maximum number of predictors to be plotted. Default is 30}

\item{max_char_length}{The maximum character length of a predictor to be printed.
Default is 5. For long variable names one may adjust this number.}
}
\value{
list of two dataframes. varimp contains the variables, group structure and
variable importance on both group and individual variable basis.
group_importance contains the the aggregated relative importance
(relative reduction of loss-function) of all group baselearners and of all
individual variables
}
\description{
Returns the variable importance of a sparse-group mboost model in a dataframe.
}
\examples{
\dontrun{
library(mboost)
library(dplyr)
set.seed(1)
df <- data.frame(
 x1 = rnorm(100),x2 = rnorm(100),x3 = rnorm(100),
 x4 = rnorm(100), x5 = runif(100)
 )
df <- df \%>\%
mutate_all(function(x){as.numeric(scale(x))})
df$y <- df$x1+df$x4+df$x5
group_df <- data.frame(
 group_name = c(1,1,1,2,2),
 var_name = c('x1','x2','x3','x4','x5')
)

sgb_formula <- as.formula(create_formula(alpha = 0.3, group_df = group_df))
sgb_model <- mboost(formula = sgb_formula, data = df)
sgb_varimp <- get_varimp(sgb_model)}
}
