% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formula.R
\name{create_formula}
\alias{create_formula}
\title{Create a sparse-group boosting formula}
\usage{
create_formula(
  alpha = 0.05,
  group_df = NULL,
  blearner = "bols",
  outcome_name = "y",
  group_name = "group_name",
  var_name = "var_name",
  intercept = FALSE
)
}
\arguments{
\item{alpha}{Numeric mixing parameter. For alpha = 0 only group baselearners and for
alpha = 1 only individual baselearners are defined.}

\item{group_df}{Dataframe containing variable names with group structure}

\item{blearner}{Type of baselearner. Default is \code{'bols'}}

\item{outcome_name}{Name of dependent variable}

\item{group_name}{Name of column in group_df indicating the group structure of the variables.}

\item{var_name}{Name of column in group_df containing the variable names
to be used as predictors.}

\item{intercept}{Logical, should intercept be used?}
}
\value{
Character containing the formula to be passed to \code{\link[mboost:gamboost]{mboost::mboost()}} yielding the sparse group boosting for a given value mixing parameter alpha.
}
\description{
Creates a \code{mboost} formula that allows to fit a sparse group boosting model based on
boosted Ridge Regression with mixing parameter \code{alpha}. The formula consists of a
group baselearner part with degrees of freedom
1-alpha and individual baselearners with degrees of freedom \code{alpha}.
}
